requirements.txt
ollama>=0.1.7
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.3
langchain-core>=0.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=9.0.0
numpy>=1.21.0
